move squint core to src
