+++ resources/squint/core.edn
@@ -6,6 +6,7 @@
   _
   _PLUS_
   _iterator
+  alength
   apply
   array_QMARK_
   assoc
